+++ examples/shopping-list/config.app.js
@@ -27,7 +27,9 @@ const appConfig = {
       path: '/api/shopping-lists.json',
       handler: 'common/api-handlers/shopping_lists.js',
       method: 'GET'
-    },
+    }
+    /*
+    ,
     {
       handler: {
         module: 'resolve-runtime/lib/common/handlers/live-require-handler.js',
@@ -39,9 +41,12 @@ const appConfig = {
       path: '/:markup*',
       method: 'GET'
     }
+     */
   ],
   clientEntries: [
-    'client/index.js',
+    'client/index.js'
+    /*
+    ,
     [
       'client/ssr.js',
       {
@@ -58,6 +63,7 @@ const appConfig = {
         target: 'node'
       }
     ]
+    */
   ]
 }