+++ packages/core/resolve-redux/src/command/actions.ts
@@ -1,71 +1,50 @@
+import { Command } from 'resolve-client'
 import {
   SEND_COMMAND_FAILURE,
   SEND_COMMAND_REQUEST,
   SEND_COMMAND_SUCCESS
 } from '../action-types'
 
+export type CommandAction = {
+  type: string
+  command: Command
+}
+
 export type SendCommandRequestAction = {
   type: typeof SEND_COMMAND_REQUEST
-  commandType: string
-  aggregateId: string
-  aggregateName: string
-  payload: any
-}
+  usedByHook: boolean
+} & CommandAction
 export const sendCommandRequest = (
-  commandType: string,
-  aggregateId: string,
-  aggregateName: string,
-  payload: any
+  command: Command,
+  usedByHook: boolean
 ): SendCommandRequestAction => ({
   type: SEND_COMMAND_REQUEST,
-  commandType,
-  aggregateId,
-  aggregateName,
-  payload
+  command,
+  usedByHook
 })
 
 export type SendCommandSuccessAction = {
   type: typeof SEND_COMMAND_SUCCESS
-  commandType: string
-  aggregateId: string
-  aggregateName: string
-  payload: any
   result: any
-}
+} & CommandAction
 export const sendCommandSuccess = (
-  commandType: string,
-  aggregateId: string,
-  aggregateName: string,
-  payload: any,
+  command: Command,
   result: any
 ): SendCommandSuccessAction => ({
   type: SEND_COMMAND_SUCCESS,
-  commandType,
-  aggregateId,
-  aggregateName,
-  payload,
+  command,
   result
 })
 
 export type SendCommandFailureAction = {
   type: typeof SEND_COMMAND_FAILURE
-  commandType: string
-  aggregateId: string
-  aggregateName: string
-  payload: any
   error: any
-}
+} & CommandAction
 export const sendCommandFailure = (
-  commandType: string,
-  aggregateId: string,
-  aggregateName: string,
-  payload: any,
+  command: Command,
   error: Error
 ): SendCommandFailureAction => ({
   type: SEND_COMMAND_FAILURE,
-  commandType,
-  aggregateId,
-  aggregateName,
-  payload,
+  command,
   error
 })