+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## 0.25.0
 
+### Breaking Changes
+
+#### resolve-redux
+
+- action types moved to internal namespace, avoid using them
+- internal actions structure drastically changed, avoid using them, they can change in future
+
+
 ### Added
 
 - [Incremental import module](packages/modules/resolve-module-incremental-import/) (packages/modules/resolve-module-incremental-import/) provides out-of-the-box api-handlers that allow you to implement incremental import in any application..
@@ -15,6 +23,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `pushIncrementalImport` - accumulate events for incremental import 
   - `commitIncrementalImport` - commit accumulated events to eventstore 
   - `rollbackIncrementalImport` - drop accumulated events
+- resolve-redux
+  - new redux hook API
 
 ### Changed